Introducing: API V2
Some constraints in the current pubsub API made it a little hard to
modify and extend it to fulfil new requirements. As such, I'm
introducing a version 2 of the API, that is more RESTful and
representative of the domain.

A /v2/subscriptions route is now used to handle all subscription
requests, however v1 continues to exist, and they both create records of
the same NDB entity kind, meaning they can both continue to be used in
parallel.